requirements.txt
# Framework principal
Flask==2.3.3

# ORM y conexión con bases de datos
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
psycopg2-binary==2.9.9  # Para soporte PostgreSQL (Railway también usa esto a veces)

# Autenticación JWT
Flask-JWT-Extended==4.6.0

# Variables de entorno y utilidades
python-dotenv==1.0.1
Werkzeug==3.0.3

# Documentación interactiva de la API (Swagger)
Flasgger==0.9.7b2
PyYAML==6.0.2

# Servidor de producción (Railway / Codespaces)
gunicorn==21.2.0

# Manejo de CORS (opcional pero recomendado para frontends)
Flask-Cors==4.0.0

bcrypt==4.1.2
python-dateutil==2.8.2
